# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface TripInput {
  /** Driver's current location (city, state or address) */
  currentLocation: string;
  /** Cargo pickup location */
  pickupLocation: string;
  /** Cargo dropoff location */
  dropoffLocation: string;
  /**
     * Hours already used in the current 70-hour/8-day cycle
     * @minimum 0
     * @maximum 70
     */
  currentCycleUsed: number;
}

export interface RouteSegment {
  from: string;
  to: string;
  distanceMiles: number;
  durationHours: number;
  /**
     * @minItems 2
     * @maxItems 2
     */
  fromCoords: number[];
  /**
     * @minItems 2
     * @maxItems 2
     */
  toCoords: number[];
}

export interface RouteInfo {
  /** Total trip distance in miles */
  totalDistanceMiles: number;
  /** Total pure driving time in hours */
  totalDrivingHours: number;
  /**
     * Array of [lat, lng] coordinate pairs for the route
     * @items.minItems 2
     * @items.maxItems 2
     */
  polyline: number[][];
  segments: RouteSegment[];
}

/**
 * Type of stop
 */
export type StopType = typeof StopType[keyof typeof StopType];


export const StopType = {
  start: 'start',
  pickup: 'pickup',
  dropoff: 'dropoff',
  rest: 'rest',
  fuel: 'fuel',
  break: 'break',
} as const;

export interface Stop {
  /** Type of stop */
  type: StopType;
  /** Human-readable location name */
  location: string;
  lat: number;
  lng: number;
  /** Day number (1-based) of arrival */
  arrivalDay: number;
  /** Hour of day (0-24) of arrival */
  arrivalHour: number;
  /** Time spent at this stop in hours */
  durationHours: number;
  /** Total miles driven to reach this stop */
  cumulativeMiles: number;
  /** Reason for stop / regulatory citation */
  notes: string;
}

export type DutyStatusEntryStatus = typeof DutyStatusEntryStatus[keyof typeof DutyStatusEntryStatus];


export const DutyStatusEntryStatus = {
  offDuty: 'offDuty',
  sleeperBerth: 'sleeperBerth',
  driving: 'driving',
  onDutyNotDriving: 'onDutyNotDriving',
} as const;

export interface DutyStatusEntry {
  status: DutyStatusEntryStatus;
  /** Start time as fractional hour (0-24) */
  startHour: number;
  /** End time as fractional hour (0-24) */
  endHour: number;
  location: string;
}

export interface RecapInfo {
  totalOnDutyToday: number;
  totalOnDutyLast7Days: number;
  totalOnDutyAvailableTomorrow: number;
  cycleHoursUsed: number;
  cycleHoursRemaining: number;
}

export interface EldLog {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** 1-based day number of the trip */
  dayNumber: number;
  startLocation: string;
  endLocation: string;
  /** Miles driven today */
  totalMilesDriving: number;
  /** Total mileage odometer today */
  totalMilesToday: number;
  entries: DutyStatusEntry[];
  remarks: string;
  totalHoursOffDuty: number;
  totalHoursSleeperBerth: number;
  totalHoursDriving: number;
  totalHoursOnDutyNotDriving: number;
  recap: RecapInfo;
}

export interface TripSummary {
  totalDays: number;
  totalDrivingHours: number;
  totalOnDutyHours: number;
  totalDistanceMiles: number;
  hosCompliant: boolean;
  warnings: string[];
  fuelStops: number;
  restStops: number;
  breakStops: number;
}

export interface TripPlan {
  route: RouteInfo;
  stops: Stop[];
  eldLogs: EldLog[];
  summary: TripSummary;
}

export interface GeocodedLocation {
  displayName: string;
  lat: number;
  lng: number;
}

export type GeocodeLocationParams = {
/**
 * Location string to geocode
 */
q: string;
};