# lib/api-zod/src/generated/types/dutyStatusEntry.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { DutyStatusEntryStatus } from './dutyStatusEntryStatus';

export interface DutyStatusEntry {
  status: DutyStatusEntryStatus;
  /** Start time as fractional hour (0-24) */
  startHour: number;
  /** End time as fractional hour (0-24) */
  endHour: number;
  location: string;
}