# lib/api-zod/src/generated/types/eldLog.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { DutyStatusEntry } from './dutyStatusEntry';
import type { RecapInfo } from './recapInfo';

export interface EldLog {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** 1-based day number of the trip */
  dayNumber: number;
  startLocation: string;
  endLocation: string;
  /** Miles driven today */
  totalMilesDriving: number;
  /** Total mileage odometer today */
  totalMilesToday: number;
  entries: DutyStatusEntry[];
  remarks: string;
  totalHoursOffDuty: number;
  totalHoursSleeperBerth: number;
  totalHoursDriving: number;
  totalHoursOnDutyNotDriving: number;
  recap: RecapInfo;
}