# lib/api-zod/src/generated/types/geocodedLocation.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface GeocodedLocation {
  displayName: string;
  lat: number;
  lng: number;
}