# lib/api-zod/src/generated/types/recapInfo.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface RecapInfo {
  totalOnDutyToday: number;
  totalOnDutyLast7Days: number;
  totalOnDutyAvailableTomorrow: number;
  cycleHoursUsed: number;
  cycleHoursRemaining: number;
}