# lib/api-zod/src/generated/types/routeInfo.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteSegment } from './routeSegment';

export interface RouteInfo {
  /** Total trip distance in miles */
  totalDistanceMiles: number;
  /** Total pure driving time in hours */
  totalDrivingHours: number;
  /**
     * Array of [lat, lng] coordinate pairs for the route
     * @items.minItems 2
     * @items.maxItems 2
     */
  polyline: number[][];
  segments: RouteSegment[];
}