# lib/api-zod/src/generated/types/routeSegment.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteSegment {
  from: string;
  to: string;
  distanceMiles: number;
  durationHours: number;
  /**
     * @minItems 2
     * @maxItems 2
     */
  fromCoords: number[];
  /**
     * @minItems 2
     * @maxItems 2
     */
  toCoords: number[];
}