# lib/api-zod/src/generated/types/stop.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { StopType } from './stopType';

export interface Stop {
  /** Type of stop */
  type: StopType;
  /** Human-readable location name */
  location: string;
  lat: number;
  lng: number;
  /** Day number (1-based) of arrival */
  arrivalDay: number;
  /** Hour of day (0-24) of arrival */
  arrivalHour: number;
  /** Time spent at this stop in hours */
  durationHours: number;
  /** Total miles driven to reach this stop */
  cumulativeMiles: number;
  /** Reason for stop / regulatory citation */
  notes: string;
}