# lib/api-zod/src/generated/types/tripInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface TripInput {
  /** Driver's current location (city, state or address) */
  currentLocation: string;
  /** Cargo pickup location */
  pickupLocation: string;
  /** Cargo dropoff location */
  dropoffLocation: string;
  /**
     * Hours already used in the current 70-hour/8-day cycle
     * @minimum 0
     * @maximum 70
     */
  currentCycleUsed: number;
}