# lib/api-zod/src/generated/types/tripPlan.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { EldLog } from './eldLog';
import type { RouteInfo } from './routeInfo';
import type { Stop } from './stop';
import type { TripSummary } from './tripSummary';

export interface TripPlan {
  route: RouteInfo;
  stops: Stop[];
  eldLogs: EldLog[];
  summary: TripSummary;
}