# lib/api-zod/src/generated/types/tripSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * HOS Trip Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface TripSummary {
  totalDays: number;
  totalDrivingHours: number;
  totalOnDutyHours: number;
  totalDistanceMiles: number;
  hosCompliant: boolean;
  warnings: string[];
  fuelStops: number;
  restStops: number;
  breakStops: number;
}